pyproject.toml
[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARN"

[tool.poetry]
name = "energy-dev-case"
version = "0.1.0"
description = ""
authors = ["Guillermo Enrique Sánchez de Abreu <naselliug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = "^2.0.20"
fastapi = "0.101.0"
pydantic = "^2.3.0"
psycopg2-binary = "^2.9.5"
uvicorn = "^0.23.2"
pyyaml = "^6.0"
aiofiles = "^23.2.1"
python-multipart = "^0.0.6"
starlette-exporter = "^0.16.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"

[tool.poetry.group.test.dependencies]
pytest-mock = "^3.10.0"
pytest = "^7.2.1"
testcontainers-postgres = "^0.0.1rc1"
httpx = "^0.23.3"
coverage = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.tox]
legacy_tox_ini = """
    [tox]
    env_list =
        py312
        py311
        py310
        py39
        py38

    [testenv]
    skip_install = true
    allowlist_externals = poetry
"""
